requirements.txt
streamlit==1.28.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0
evaluate>=0.4.0
rouge-score>=0.1.2
pandas>=1.5.0
numpy>=1.24.0
langdetect>=1.0.9
openai-whisper>=20230314
moviepy>=1.0.3
pdfplumber>=0.9.0
python-docx>=0.8.11
psutil>=5.9.0
concurrent-futures>=3.1.1
python-multipart>=0.0.6
pillow>=9.0.0
scipy>=1.10.0
librosa>=0.10.0
soundfile>=0.12.0
